feat: Implement a new multi-step booking flow with dedicated components and composable,, and enhance appointment store availability calculation with safety checks.
